+++ CHANGELOG.md
@@ -2,6 +2,14 @@ GAMSPy CHANGELOG
 ================
 
 -------------------------------------------------------------------------------
+GAMSPy 0.11.4
+-------------------------------------------------------------------------------
+- General
+  - Expose GTP special values via gp.SpecialValues
+- Testing
+  - Test the special value usage in assignments for Parameter, ImplicitParameter and Operation (Sum, Smax, Smin, Product).
+
+-------------------------------------------------------------------------------
 GAMSPy 0.11.3
 -------------------------------------------------------------------------------
 - General